# optimize query pack